CameraGroup: Allow multiple conditional compositors
Allow multiple conditional compositor paths to be specified per camera,
with the compositor with the first met condition being chosen (falling
back to the default compositor path).

Multiple <compositor> nodes can be specified. They can be of either the
existing unconditional form:
  <compositor>Compositor/default</compositor>

or conditional on some property expression such as this:
  <compositor>
    <path>Compositor/...</path>
    <condition>...</condition>
  <compositor>

This allows the entire compositor to be changed depending on some
property expression, for example to switch the main screen between a
normal scene view and a mirror of a VR view on a compositor reload based
on settings.