CameraGroup: Add VR mirror setup
Set up the default camera as a mirror of the VR view when needed, using
a new CameraInfo flag named VR_MIRROR based on a camera property
"vr-mirror" which is set on the default camera group. When set this flag
causes the compositor to be constructed using buildVRMirrorCompositor()
rather than the usual Compositor::create().