Add new VR input subsystem
Add a new input subsystem called FGVRInput for VR inputs from osgXR /
OpenXR. This requires a slightly newer version of osgXR.

The subsystem is only built when VR support is enabled, and is
configured via the property tree, specifically nodes inside /input/vr/.
It extends the osgXR action classes and constructs them from the XML
definitions.

It also adds a further abstraction of "modes" which can be attached to
subactions (i.e. hands) as a stack, and an action set, to interpret the
inputs in a particular way according to bindings defined in XML.