FGVRInput: Duplicate mode property tree per subaction
Duplicate entire interaction mode property tree for each subaction so as
to allow separate bindings to function propertly.

The existing isolation code isn't as effective as originally hoped.
Although each subaction/mode has a nasal module, they share binding
nodes so the bindings would have to be tweaked to give indication of
which subaction is active in order for generic fgcommands to work. That
isn't practical due to the abstraction of bindings via FGButton and
FGCommonInput.