Add fgcommands vr-mode-push & vr-mode-toggle
Add a couple of fgcommands for controlling the VR interaction mode
stack:

The vr-mode-push command pushes an interaction mode onto the stack above
the current mode (replacing any other items in front of it on the
stack).

The vr-mode-toggle command removes any modes in front of the current
mode from the stack, or if there are no such modes it pushes the chosen
interaction mode onto the stack.

The current mode is determined by looking up the tree for a mode node,
and so assumes it will be called from a binding within a mode.