FGVRInput: Cope with getStringValue() returning std::string
Fix FGVRInput to cope with SGPropertyNode::getStringValue() returning
std::string, and the removal of SGPropertyNode::getName() in favour of
getNameString().

This involves the conversion of const char pointers to std::string
(passing to functions by reference), the use of std::string::empty()
instead of pointer validation, and the use of std::string comparison
rather than strcmp().