Add FGVRHand
Add an FGVRHand class, based on osgXR::HandPose, which modifies a
tracked hand pose to interact with the scene.

The plane of motion of each finger is intersected with the scene to
obtain cross sections of the scene which the finger may interract with.
The new FGVRCollision code is then used to determine the ranges of
finger squeeze (from outstretched to clenched fist) which intersect the
scene. This is then used to move the fingers to avoid intersection with
the scene, and to squeeze in to touch the surfaces in the scene ready to
interact with them.

Though the fingers avoid intersection, the palm does not yet collide,
and nor do stretched back fingers. Nor is interaction with controls yet
implemented.